Rendering is fixed.
The image does now render, it is not rendering properly when running an APNG file. This is because there is no handling of blend operations.

The way the original wxAnimationCtrl code is written the rendering of a frame is done by wxAnimationControl. This design is not going to work correctly for APNG files because of the blend operations that need to be performed when rendering the frame.

I am going to make the changes needed to correct this issue.