Moved DrawFrame to wxAnimationDecoder. Added limited support for FLIF
I moved the rendering of the frame to the decoder because each image format may have it's own special way of rendering the image. In the case of APNG there are blend operations that dictate how the frame gets rendered. These types of operations are not accounted for in the AnimationCtrl class.

I added support for Free Lossless Image Format (FLIF). FLIF does support animations and I thought it should be added. In order to use FLIF a Python extension module needed to be made. At the moment I only have an extension module compiled for Python 2.7 x86 and Python 3.5+ x64. Both extension modules are for Microsoft Windows. I will add extension modules for Windows, Darwin and Debian supporting Python 2.7, 3.5+ x86 and x64. This will not be until I write a setup program and also set up a CI on Azure. I am seeing how many additional file formats I am going to be able to add.

I plan on adding thee file formats WebP, WebM, BPG, MNG and AVIF