Started coding the build system.
This is a complex build system because of the need to compile several static libraries where some of them have a dependency to another and also because we are going to be compiling using cffi instead of distutils or setup tools.

 I am retrofitting a build system I had started on several months ago. I am attempting to write a build framework at the same time. Something that I will be able to use between different projects.

The build framework is going to have support for the following

building documentation
building wheels
building eggs
installing
compiling static libraries
compiling dynamic libraries
compiling cffi extensions
compiling c extensions
compiling cython extensions

on some of the compiling tasks there is the ability to optionally use all cores on the system. It does this by using (n) threads + 1 where n is the number of cores in the system. the 1 is the main thread that manages all of the other threads. Instead of passing all of the file names to be compiled at once I cut the list of files up and pass equal numbers of them to each of the threads. each thread then runs the compiler for each file using subprocess. The system automatically place the new compiler process to run on a core that does not already have a compiler process running on it. So instead of using the subprocess module which would be a mess to make work correctly, I opted to let the users OS handle what core the process is going to run on. It made it easy to do.

The following platforms will be supported (for now)
Cygwin
OSX
Linux (Debian(ish) flavors)
SunOS
Windows
FreeBSD

I will be expanding the build system to add more functionality and to add additional platforms.
My main goal is to make a build system that works correctly while being easy to use. I am not there as of yet I still have to do a whole lot more in the ease of use department. I am getting there. The biggest road block is Windows. I have most of that out of the way. I do have to add onto the msvc portion so it handles Visual Studio 2019 better then it currently does.